+++ Engine/Source/Runtime/Core/Private/Android/AndroidPlatformFile.cpp
@@ -64,9 +64,9 @@ namespace
 		}
 
 		return FFileStatData(
-			AndroidEpoch + FTimespan::FromSeconds(FileInfo.st_ctime), 
-			AndroidEpoch + FTimespan::FromSeconds(FileInfo.st_atime), 
-			AndroidEpoch + FTimespan::FromSeconds(FileInfo.st_mtime), 
+			AndroidEpoch + FTimespan::FromSeconds((double)FileInfo.st_ctime),
+			AndroidEpoch + FTimespan::FromSeconds((double)FileInfo.st_atime),
+			AndroidEpoch + FTimespan::FromSeconds((double)FileInfo.st_mtime), 
 			FileSize,
 			bIsDirectory,
 			!(FileInfo.st_mode & S_IWUSR)
