+++ Engine/Source/Runtime/Core/Private/Android/AndroidPlatformMisc.cpp
@@ -916,7 +916,7 @@ FAndroidMisc::FCPUState& FAndroidMisc::GetCPUState(){
 		}
 		fclose(FileHandle);
 
-		double WallTime;
+		uint64_t WallTime;
 		double CPULoad[CurrentCPUState.CoreCount];
 		CurrentCPUState.AverageUtilization = 0.0;
 		for (size_t n = 0; n < CurrentCPUState.CoreCount; n++) {
@@ -932,7 +932,7 @@ FAndroidMisc::FCPUState& FAndroidMisc::GetCPUState(){
 				CPULoad[n] = 0;
 				continue;
 			}
-			CPULoad[n] = (WallTime - (double)IdleTime) * 100.0 / WallTime;
+			CPULoad[n] = ((double)WallTime - (double)IdleTime) * 100.0 / (double)WallTime;
 			CurrentCPUState.Utilization[n] = CPULoad[n];
 			CurrentCPUState.AverageUtilization += CPULoad[n];
 		}
